backend/requirements.txt
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
django-cors-headers>=4.4,<5.0
django-filter>=24.0,<25.0
mysqlclient>=2.2,<3.0
Pillow>=10.0,<11.0
redis>=5.0,<6.0
celery>=5.4,<6.0
channels>=4.1,<5.0
channels-redis>=4.2,<5.0
paho-mqtt>=2.1,<3.0
gunicorn>=22.0,<23.0
